split tests into two files
Makes the testing setup a bit more flexible.
